+++ src/modules/auth.configuration.ts
@@ -16,7 +16,7 @@ export class DefaultConfiguration {
     start_checksession = false;
     silent_renew = true;
     silent_renew_offset_in_seconds = 0;
-    startup_route = '/dataeventrecords';
+    post_login_route = '/';
     // HTTP 403
     forbidden_route = '/forbidden';
     // HTTP 401
@@ -50,7 +50,7 @@ export class OpenIDImplicitFlowConfiguration {
     start_checksession: boolean;
     silent_renew: boolean;
     silent_renew_offset_in_seconds: number;
-    startup_route: string;
+    post_login_route: string;
     forbidden_route: string;
     unauthorized_route: string;
     auto_userinfo: boolean;
@@ -112,8 +112,8 @@ export class AuthConfiguration {
         return this.openIDImplicitFlowConfiguration.silent_renew_offset_in_seconds || this.defaultConfig.silent_renew_offset_in_seconds;
     }
 
-    get startup_route(): string {
-        return this.openIDImplicitFlowConfiguration.startup_route || this.defaultConfig.startup_route;
+    get post_login_route(): string {
+        return this.openIDImplicitFlowConfiguration.post_login_route || this.defaultConfig.post_login_route;
     }
 
     get forbidden_route(): string {